Move NotificationsService to notifications directory
