+++ src/app/app.module.ts
@@ -29,7 +29,7 @@ import { TeamsService } from './teams/teams.service';
 import { RatingService } from './shared/forms/rating.service';
 import { TagsService } from './shared/forms/tags.service';
 import { ManagerService } from './manager-dashboard/manager.service';
-import { NotificationsService } from './shared/notifications.service';
+import { NotificationsService } from './notifications/notifications.service';
 
 @NgModule({
   imports: [
